requirements.txt
# requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-generativeai==0.7.0
pydantic==2.7.4
python-dotenv==1.0.1 # Optional, for local development to load .env files
